requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
pillow==10.2.0
numpy==1.26.3
loguru==0.7.2
opencv-python==4.9.0.80